--- main.py
@@ -1,7 +1,0 @@
-print("Hello World!")
-
-print("Damian")
-
-print("Test")
-
-print("Test2")
